Rename float to prevent overriding system header Float.h -> RNFloat.h
